+++ app/src/main/java/com/badcompany/pitakpass/domain/repository/DriverPostRepository.kt
@@ -9,7 +9,7 @@ import com.badcompany.pitakpass.util.ResultWrapper
 interface DriverPostRepository {
 
 //    suspend fun filterDriverPost(filter: Filter): ResultWrapper<List<DriverPost>>
-    suspend fun getDriverPostById(id: Int): ResponseWrapper<DriverPost>
+    suspend fun getDriverPostById(id: Long): ResponseWrapper<DriverPost>
     suspend fun joinARide(myOffer: PassengerOffer): ResponseWrapper<Any>
